+++ AWSCore/Service/AWSServiceEnum.h
@@ -135,6 +135,10 @@ typedef NS_ENUM(NSInteger, AWSRegionType) {
      *  Europe (Milan)
      */
     AWSRegionEUSouth1 NS_SWIFT_NAME(EUSouth1),
+    /**
+     *  Europe (Spain)
+     */
+    AWSRegionEUSouth2 NS_SWIFT_NAME(EUSouth2),
 };
 
 /**